services/pyproject.toml
[tool.poetry]
name = "dr-lingo-services"
version = "0.1.0"
description = "Django REST API services for dr-lingo projects"
authors = ["Dev Shack Academy"]
packages = [{include = "api"}, {include = "config"}]

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.3.0"
django-cors-headers = "^4.0.0"
psycopg2-binary = "^2.9.0"
python-decouple = "^3.8"
dj-database-url = "^2.1.0"
drf-yasg = "^1.21.7"
google-generativeai = "^0.3.0"
pillow = "^10.0.0"
python-magic = "^0.4.27"
whitenoise = "^6.6.0"
# Production Infrastructure
celery = {extras = ["redis"], version = "^5.3.0"}
django-redis = "^5.4.0"
redis = "^5.0.0"
pika = "^1.3.0"
django-celery-results = "^2.5.0"
django-celery-beat = "^2.5.0"
gunicorn = "^21.2.0"
# WebSocket Support
channels = "^4.0.0"
channels-redis = "^4.2.0"
daphne = "^4.1.0"
# Monitoring
flower = "^2.0.0"
# Hugging Face datasets (no audio extra - we disable audio decoding)
datasets = "^3.6.0"
huggingface-hub = "^0.25.0"
# Two-Factor Authentication (OTP)
django-otp = "^1.5.0"
django-two-factor-auth = {extras = ["email"], version = "^1.16.0"}
qrcode = "^7.4.2"
piper-tts = "^1.3.0"
# PDF Processing
pypdf = "^4.0.0"
pdf2image = "^1.17.0"
pytesseract = "^0.3.13"
# Google Cloud Storage for media files
django-storages = {extras = ["google"], version = "^1.14.0"}
google-cloud-storage = "^2.10.0"
google-cloud-pubsub = "^2.18.0"


[tool.poetry.group.dev.dependencies]
pre-commit = "^4.3.0"
pytest = "^8.4.2"
pytest-django = "^4.8.0"
pytest-cov = "^6.0.0"
pytest-asyncio = "^0.24.0"
ipython = "^9.5.0"

[build-system]
requires = ["poetry-core"]
build-services = "poetry.core.masonry.api"

[tool.ruff]
line-length = 119
indent-width = 4
target-version = "py311"
src = ["api", "config"]

[tool.ruff.lint]
select = ["E4", "E7", "E9", "F", "I001"]
fixable = ["ALL"]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["E402"]
"**/migrations/*" = ["E501"]
"**/tests/*" = ["E402"]

[tool.ruff.lint.isort]
case-sensitive = true

[tool.ruff.lint.isort.sections]
"django" = ["django"]

[tool.pytest.ini_options]
minversion = "6.0"
DJANGO_SETTINGS_MODULE = "config.settings"
pythonpath = ["."]
testpaths = ["api/tests"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
